+++ test/decorators-error/playwright.config.ts
@@ -2,8 +2,10 @@ import { defineConfig } from '@playwright/test';
 import { defineBddConfig } from 'playwright-bdd';
 
 const testDir = defineBddConfig({
-  importTestFrom: 'fixtures/index.ts',
   paths: [`features/${process.env.FEATURE}`],
+  importTestFrom: '../decorators/steps/fixtures.ts',
+  require: ['../decorators/steps/steps.ts'],
+  statefulPoms: true,
 });
 
 export default defineConfig({