+++ test/decorators/playwright.config.ts
@@ -2,8 +2,9 @@ import { defineConfig } from '@playwright/test';
 import { defineBddConfig } from 'playwright-bdd';
 
 const testDir = defineBddConfig({
-  importTestFrom: 'steps/fixtures.ts',
   paths: ['features/*.feature'],
+  importTestFrom: 'steps/fixtures.ts',
+  require: ['steps/steps.ts'],
 });
 
 export default defineConfig({